+++ span_marker/__init__.py
@@ -1,5 +1,7 @@
 __version__ = "1.0.0.dev"
 
+import logging
+
 from transformers import AutoConfig, AutoModel, TrainingArguments
 
 from span_marker.configuration import SpanMarkerConfig
@@ -8,3 +10,6 @@ from span_marker.trainer import Trainer
 
 AutoConfig.register("span-marker", SpanMarkerConfig)
 AutoModel.register(SpanMarkerConfig, SpanMarkerModel)
+
+logger = logging.getLogger("span_marker")
+logger.setLevel(logging.INFO)